+++ producer/src/main/java/com/owlexpress/producer/common/constant/ProductType.java
@@ -1,12 +1,13 @@
-package com.owlexpress.producer.domain.entity.constant;
+package com.owlexpress.producer.common.constant;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum CompanyType {
-    PRODUCER("Producer");
+public enum ProductType {
+
+    NORMAL("NORMAL"), FRESH("FRESH");
 
     private final String value;
 }
