+++ producer/src/main/java/com/owlexpress/producer/presentation/ProducerUsecase.java
@@ -1,0 +1,34 @@
+package com.owlexpress.producer.presentation;
+
+import com.owlexpress.producer.common.dto.request.CreateProducerRequestDto;
+import com.owlexpress.producer.common.dto.request.UpdateProducerRequestDto;
+import com.owlexpress.producer.domain.entity.Producer;
+import java.util.UUID;
+import org.springframework.transaction.annotation.Transactional;
+
+public interface ProducerUsecase {
+
+    @Transactional
+    void create(
+            CreateProducerRequestDto createProducerRequestDto,
+            String passport
+    );
+
+    @Transactional
+    void update(
+            UpdateProducerRequestDto updateProducerRequestDto,
+            UUID producerId,
+            String passport
+    );
+
+    void updateProducer(
+            Producer producer,
+            UpdateProducerRequestDto dto
+    );
+
+    @Transactional
+    void delete(
+            UUID producerId,
+            String passport
+    );
+}
