+++ producer/src/main/java/com/owlexpress/producer/presentation/dto/request/CreateProductInfoRequestDto.java
@@ -2,7 +2,7 @@ package com.owlexpress.producer.presentation.dto.request;
 
 import com.owlexpress.producer.domain.entity.Producer;
 import com.owlexpress.producer.domain.entity.ProductInfo;
-import com.owlexpress.producer.domain.entity.constant.ProductType;
+import com.owlexpress.producer.common.constant.ProductType;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
